+++ src/schema.rs
@@ -4,7 +4,7 @@ diesel::table! {
     highscore (id) {
         id -> Integer,
         name -> Text,
-        score -> Float,
+        score -> Integer,
         created_at -> Nullable<Timestamp>,
     }
 }
